scylla_cql/serialize: Extract value tests to a separate file
